starting over with testing
